changed the double types to BigDecimals in tests

+++ core/src/test/java/dev/andrylat/raqimbek/bankingutils/core/services/mortgagecalculator/MonthlyMortgagePaymentTestData.java
@@ -1,7 +1,9 @@
 package dev.andrylat.raqimbek.bankingutils.core.services.mortgagecalculator;
 
+import java.math.BigDecimal;
+
 public record MonthlyMortgagePaymentTestData(
-    double borrowedAmount,
-    double annualInterestRate,
-    double numberOfYears,
-    long expectedMonthlyPayment) {}
+    BigDecimal borrowedAmount,
+    BigDecimal annualInterestRate,
+    BigDecimal numberOfYears,
+    BigDecimal expectedMonthlyPayment) {}
